requirements.txt
numpy>=1.24.0
python-dotenv>=1.0.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-groq>=0.1.0
sentence-transformers>=3.0.0
groq>=0.9.0
chromadb>=0.5.0
langchain-huggingface>=0.3.0
langchain-core>=0.3.0
streamlit>=1.39.0
langchain-chroma>=0.1.0